autopilot+test: make centrality test data available for other tests
